backend/requirements.txt
asgiref==3.9.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Brotli==1.1.0
click==8.3.0
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
huepy==1.2.1
Jinja2==3.1.6
livereload==2.7.1
markdown2==2.5.4
MarkupSafe==3.0.2
npm==0.1.1
optional-django==0.1.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
tornado==6.5.2
uvicorn==0.37.0
vite==1.5.2
whitenoise==6.11.0
